docs for create_info_dicts, move out to msv4_info_dics, #247
